feat: 쿠폰 낙관적 락 기반 동시성 제어 구현

+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/coupon/IssuedCouponJpaRepository.java
@@ -3,5 +3,5 @@ package com.loopers.infrastructure.coupon;
 import com.loopers.domain.coupon.IssuedCoupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IssuedCouponJpaRepository extends JpaRepository<IssuedCoupon, Integer> {
+public interface IssuedCouponJpaRepository extends JpaRepository<IssuedCoupon, Long> {
 }
